Move quick info tests to new framework
